requirements.txt
pika
faker
python-dotenv
